refactor(sync): extract sync-server into standalone package
Promote the @epicenter/sync/server subpath export into its own
@epicenter/sync-server workspace package, completing the three-layer
split: sync (protocol) / sync-client (providers) / sync-server (handlers).

- Move handlers.ts from sync → sync-server, update internal imports
- Remove ./server subpath and Awareness re-export from @epicenter/sync
- Update server-local and server-remote-cloudflare deps and imports
- Mark spec as implemented
