+++ packages/sync-server/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@epicenter/sync",
+	"name": "@epicenter/sync-server",
 	"version": "0.0.1",
 	"main": "./src/index.ts",
 	"types": "./src/index.ts",
@@ -11,6 +11,7 @@
 		"typecheck": "tsc --noEmit"
 	},
 	"dependencies": {
+		"@epicenter/sync": "workspace:*",
 		"lib0": "catalog:",
 		"y-protocols": "catalog:"
 	},
